+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+- Changed [pawk] Use subexpression operator to expand column specifications inside double quotes
 - Changed [Invoke-Link] Tag management
+- Translated [README.md] synopsis of `man2` function Japanese to English: `man2` function
 
 ## [7.2.1] - 2024-03-17